BLOG_POST/requirement.txt
colorama==0.4.6
mysql-connector-python==8.0.28
